requirements.txt
annotated-types==0.7.0
anyio==4.6.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
httptools==0.6.1
idna==3.10
iniconfig==2.0.0
marshmallow==3.22.0
marshmallow_dataclass==8.7.1
mypy-extensions==1.0.0
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.1
typeguard==4.3.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
